requirements.txt
Appium-Python-Client==0.28
atomicwrites==1.2.1
attrs==18.2.0
funcsigs==1.0.2
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
py==1.6.0
pytest==3.8.0
scandir==1.9.0
selenium==3.14.0
six==1.11.0
urllib3==1.23
